Move display logic from song_server to display_server
